requirements/base.txt
BeautifulSoup==3.2.1
Django==1.4.3
Fabric==1.4.3
South==0.7.6
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.17
boto==2.6.0
celery==3.0.11
distribute==0.6.28
dj-database-url==0.2.1
django-admin-bootstrapped==0.2
django-analytical==0.14.0
django-annoying==0.7.6
django-bootstrap-toolkit==2.5.9
django-celery==3.0.10
django-htmlmin==0.5.2
django-pipeline==1.2.23
django-robots==0.8.1
django-social-auth==0.7.14
django-storages==1.1.5
gevent==0.13.8
greenlet==0.4.0
gunicorn==0.16.1
httplib2==0.7.6
kombu==2.4.7
oauth2==1.5.211
psycopg2==2.4.6
pycrypto==2.6
python-dateutil==1.5
python-openid==2.2.5
raven==2.0.12.2
ssh==1.7.14
wsgiref==0.1.2
